Lidar page  update
I tried my best to clean it up and include an update, but some of my syntax may be flawed.  

Logos placement  most likely needs to be updated. 

Embedded maps need to be updated. 

I tried to delete out anything that is not needed any longer, but i may have missed some things.